requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
PyPDF2==3.0.1
pytesseract==0.3.8
pdf2image==1.16.0
Pillow>=9.0.0
reportlab==3.6.12
PyMuPDF==1.18.14
pytest==7.0.1
pytest-cov==3.0.0
coverage==6.3.2 